Move logic into GemHelper
