+++ lib/tapioca/gemfile.rb
@@ -134,6 +134,7 @@ module Tapioca
 
     class GemSpec
       extend(T::Sig)
+      include GemHelper
 
       IGNORED_GEMS = T.let(
         [
@@ -161,7 +162,7 @@ module Tapioca
 
       sig { params(gemfile_dir: String).returns(T::Boolean) }
       def ignore?(gemfile_dir)
-        gem_ignored? || gem_in_app_dir?(gemfile_dir)
+        gem_ignored? || gem_in_app_dir?(gemfile_dir, full_gem_path)
       end
 
       sig { returns(String) }
@@ -295,16 +296,6 @@ module Tapioca
       sig { returns(T::Boolean) }
       def gem_ignored?
         IGNORED_GEMS.include?(name)
-      end
-
-      sig { params(gemfile_dir: String).returns(T::Boolean) }
-      def gem_in_app_dir?(gemfile_dir)
-        !gem_in_bundle_path? && full_gem_path.start_with?(gemfile_dir)
-      end
-
-      sig { returns(T::Boolean) }
-      def gem_in_bundle_path?
-        full_gem_path.start_with?(Bundler.bundle_path.to_s, Bundler.app_cache.to_s)
       end
     end
   end
