+++ lib/tapioca/runtime/loader.rb
@@ -5,6 +5,7 @@ module Tapioca
   module Runtime
     class Loader
       extend(T::Sig)
+      include Tapioca::GemHelper
 
       sig do
         params(gemfile: Tapioca::Gemfile, initialize_file: T.nilable(String), require_file: T.nilable(String)).void
@@ -123,16 +124,6 @@ module Tapioca
             nil
           end
         end
-      end
-
-      sig { params(gemfile_dir: String, full_gem_path: String).returns(T::Boolean) }
-      def gem_in_app_dir?(gemfile_dir, full_gem_path)
-        !gem_in_bundle_path?(full_gem_path) && full_gem_path.start_with?(gemfile_dir)
-      end
-
-      sig { params(full_gem_path: String).returns(T::Boolean) }
-      def gem_in_bundle_path?(full_gem_path)
-        full_gem_path.start_with?(Bundler.bundle_path.to_s, Bundler.app_cache.to_s)
       end
     end
   end
